BooksServiceImpl refactored and tested, changed some arguments from int to long

+++ src/main/java/app/data/BooksRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
  * Created by Gabriel on 04.04.2017.
  */
 public interface BooksRepository {
-    Book retrieveBook(int index);
+    Book retrieveBook(long index);
     List<Book> retrieveBooks(String booksCategory, int booksLimit, int booksOffset);
     List<Book> retrieveNewBooks(int booksLimit, int booksOffset);
     List<String> retrieveCategoriesNames();
